src/requirements.txt
selenium
supabase
python-dotenv
pandas